requirements.txt
python-chess>=1.999
tqdm>=4.0.0
